+++ test/specs/options/rotateOnTouch.spec.js
@@ -48,7 +48,7 @@ describe('rotateOnTouch (option)', () => {
     viewer.show();
   });
 
-  it('should restore a zero-degree image after rotating within 45 degrees', (done) => {
+  it('should restore a zero-degree image after rotating 44 degrees', (done) => {
     const image = window.createImage();
     let viewer;
 
@@ -68,10 +68,10 @@ describe('rotateOnTouch (option)', () => {
       viewer.image.dispatchEvent(window.createEvent('pointermove', {
         pointerId: 2,
         pointerType: 'touch',
-        pageX: 19.92,
-        pageY: 1.74,
+        pageX: 14.39,
+        pageY: 13.89,
       }));
-      expect(viewer.imageData.rotate).to.be.closeTo(5, 0.1);
+      expect(viewer.imageData.rotate).to.be.closeTo(44, 0.1);
       viewer.image.dispatchEvent(window.createEvent('pointerup', {
         pointerId: 1,
         pointerType: 'touch',
@@ -81,6 +81,47 @@ describe('rotateOnTouch (option)', () => {
         pointerType: 'touch',
       }));
       expect(viewer.imageData.rotate).to.equal(0);
+      viewer.hide(true);
+      done();
+    });
+
+    viewer = new Viewer(image);
+    viewer.show();
+  });
+
+  it('should snap a zero-degree image to 90 degrees after rotating 46 degrees', (done) => {
+    const image = window.createImage();
+    let viewer;
+
+    image.addEventListener('viewed', () => {
+      viewer.image.dispatchEvent(window.createEvent('pointerdown', {
+        pointerId: 1,
+        pointerType: 'touch',
+        pageX: 0,
+        pageY: 0,
+      }));
+      viewer.image.dispatchEvent(window.createEvent('pointerdown', {
+        pointerId: 2,
+        pointerType: 'touch',
+        pageX: 20,
+        pageY: 0,
+      }));
+      viewer.image.dispatchEvent(window.createEvent('pointermove', {
+        pointerId: 2,
+        pointerType: 'touch',
+        pageX: 13.89,
+        pageY: 14.39,
+      }));
+      expect(viewer.imageData.rotate).to.be.closeTo(46, 0.1);
+      viewer.image.dispatchEvent(window.createEvent('pointerup', {
+        pointerId: 1,
+        pointerType: 'touch',
+      }));
+      viewer.image.dispatchEvent(window.createEvent('pointerup', {
+        pointerId: 2,
+        pointerType: 'touch',
+      }));
+      expect(viewer.imageData.rotate).to.equal(90);
       viewer.hide(true);
       done();
     });
@@ -123,7 +164,7 @@ describe('rotateOnTouch (option)', () => {
         pointerId: 2,
         pointerType: 'touch',
       }));
-      expect(viewer.imageData.rotate).to.be.closeTo(20, 0.1);
+      expect(viewer.imageData.rotate).to.equal(0);
       viewer.hide(true);
       done();
     });